requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
dj-database-url
gunicorn
Pillow
django-cors-headers==4.0.0
